activityfeed/requirements.txt
Django==1.11.10
djangorestframework==3.8.2
